Update get_candidate_codes.R
Hey Ed, I notice all the domains type start with capital letters and if you put the same word with lower case and it won't work. I added a little function to convert the 1st letter of the word to capital hence the user won't have to worry about capital letters.